datalive/requirements.txt
MySQL-python==1.2.5
Django==1.11.2
django-rest-auth==0.9.0
django-timezone-field==2.0
djangorestframework==3.5.4
djangorestframework-jwt==1.11.0
django-cors-headers
PyJWT==1.5.2
pytz==2016.10
googlemaps
oauth2client
#google-cloud==0.27.0
google-cloud-storage>=0.22.0
django-storages
django-mysql==2.2.0
openpyxl==2.2.6

## requests-toolbelt - used to monkeypatch the requests to https urls (e.g. GCS)
# https://cloud.google.com/appengine/docs/standard/python/issue-requests
requests-toolbelt

# for use with SMS/Voice
twilio
